+++ external/model-preparation-algorithm/tests/api_tests/helpers.py
@@ -3,33 +3,12 @@
 #
 
 import time
-from typing import List, Type
 
 from mpa_tasks.apis import BaseTask
 from ote_sdk.entities.datasets import DatasetEntity
 from ote_sdk.entities.metrics import Performance
 from ote_sdk.entities.model import ModelEntity
 from ote_sdk.entities.resultset import ResultSetEntity
-from ote_sdk.test_suite.training_tests_actions import (
-    BaseOTETestAction,
-    OTETestExportAction,
-    OTETestExportEvaluationAction,
-    OTETestPotAction,
-    OTETestPotEvaluationAction,
-    OTETestTrainingAction,
-    OTETestTrainingEvaluationAction,
-)
-
-
-def get_test_action_classes() -> List[Type[BaseOTETestAction]]:
-    return [
-        OTETestTrainingAction,
-        OTETestTrainingEvaluationAction,
-        OTETestExportAction,
-        OTETestExportEvaluationAction,
-        OTETestPotAction,
-        OTETestPotEvaluationAction,
-    ]
 
 
 def eval(task: BaseTask, model: ModelEntity, dataset: DatasetEntity) -> Performance:
