+++ external/model-preparation-algorithm/tests/performance_tests/helpers.py
@@ -4,23 +4,58 @@
 
 import logging
 import os.path as osp
+import time
 from collections import namedtuple
 from copy import deepcopy
-from typing import List
+from typing import List, Type
 
 from detection_tasks.extension.datasets.data_utils import load_dataset_items_coco_format
-from segmentation_tasks.extension.datasets.mmdataset import load_dataset_items
+from mpa_tasks.apis import BaseTask
 from ote_sdk.entities.datasets import DatasetEntity
 from ote_sdk.entities.label import Domain
 from ote_sdk.entities.label_schema import LabelSchemaEntity
+from ote_sdk.entities.metrics import Performance
+from ote_sdk.entities.model import ModelEntity
+from ote_sdk.entities.resultset import ResultSetEntity
 from ote_sdk.entities.subset import Subset
-from ote_sdk.test_suite.training_tests_common import (
-    ROOT_PATH_KEY,
-    make_paths_be_abs,
+from ote_sdk.test_suite.training_tests_actions import (
+    BaseOTETestAction,
+    OTETestExportAction,
+    OTETestExportEvaluationAction,
+    OTETestPotAction,
+    OTETestPotEvaluationAction,
+    OTETestTrainingAction,
+    OTETestTrainingEvaluationAction,
 )
+from ote_sdk.test_suite.training_tests_common import ROOT_PATH_KEY, make_paths_be_abs
+from segmentation_tasks.extension.datasets.mmdataset import load_dataset_items
 from torchreid_tasks.utils import ClassificationDatasetAdapter
 
 logger = logging.getLogger(__name__)
+
+
+def get_test_action_classes() -> List[Type[BaseOTETestAction]]:
+    return [
+        OTETestTrainingAction,
+        OTETestTrainingEvaluationAction,
+        OTETestExportAction,
+        OTETestExportEvaluationAction,
+        OTETestPotAction,
+        OTETestPotEvaluationAction,
+    ]
+
+
+def eval(task: BaseTask, model: ModelEntity, dataset: DatasetEntity) -> Performance:
+    start_time = time.time()
+    result_dataset = task.infer(dataset.with_empty_annotations())
+    end_time = time.time()
+    print(f"{len(dataset)} analysed in {end_time - start_time} seconds")
+    result_set = ResultSetEntity(
+        model=model, ground_truth_dataset=dataset, prediction_dataset=result_dataset
+    )
+    task.evaluate(result_set)
+    assert result_set.performance is not None
+    return result_set.performance
 
 
 def DATASET_PARAMETERS_FIELDS() -> List[str]:
