+++ external/model-preparation-algorithm/tests/performance_tests/test_ote_detection_training.py
@@ -27,8 +27,8 @@ from ote_sdk.test_suite.training_tests_helper import (
     OTETrainingTestInterface,
 )
 
-from ..helpers.common import get_test_action_classes
-from ..helpers.dummy_dataset import (
+from helpers import (
+    get_test_action_classes,
     create_object_detection_dataset_and_labels_schema,
     get_dataset_params_from_dataset_definitions,
 )
